fixed error

+++ src/perm64.rs
@@ -22,7 +22,7 @@ impl Mapping for Perm64 {
     fn order(&self) -> u8 {
         for i in (0..16).rev() {
             if ((self.0 >> (4 * i)) & 0xF) != i {
-                return i + 1;
+                return i as u8 + 1;
             }
         }
         1
